Introduce SAMI Bot the 2nd, an AI Agent instead of a Chatbot
The first SAMI Bot is just a chatbot that backend will query data
and then put in the prompt. To make it automatically fetch data
or interact with the backend I need to make an Agent instead.

Thus, SAMI Bot the 2nd was born. Not only is the prompt changed, but
we also need to make a custom toolbox for it to interact with our
backend. These include:
- Get context of a tenant (unpaid bills, contracts, etc)
- CRUD Maintenance Request
- CRUD Vehicle

Gemini 2.5 Flash is still our choice of model

Signed-off-by: hmtheboy154 <buingoc67@gmail.com>